Adding standard library
Creating standardLibrary.texi. This will be implictly included for all files.
This file defines all the standard functions, and tries to define as many of
them as it can without the template language itself. This means that any other
implementation of the processor can be very minimal.

The idea is to create a language that' so lightweight, yet extensible, so having
a standard library that's defined in the language, and requires very few external
funtions is beneficial.